+++ src/Beef.Data.Cosmos/CosmosDbQuery.cs
@@ -41,9 +41,9 @@ namespace Beef.Data.Cosmos
         /// Initializes a new instance of the <see cref="CosmosDbQuery{T}"/> class.
         /// </summary>
         /// <param name="db">The <see cref="CosmosDbBase"/>.</param>
-        /// <param name="queryArgs">The <see cref="CosmosDbArgs"/>.</param>
+        /// <param name="queryArgs">The <see cref="CosmosDbArgs{T}"/>.</param>
         /// <param name="query">A function to modify the underlying <see cref="IQueryable{T}"/>.</param>
-        internal CosmosDbQuery(CosmosDbBase db, CosmosDbArgs queryArgs, Func<IQueryable<T>, IQueryable<T>> query = null)
+        internal CosmosDbQuery(CosmosDbBase db, CosmosDbArgs<T> queryArgs, Func<IQueryable<T>, IQueryable<T>> query = null)
         {
             _db = Check.NotNull(db, nameof(db));
             QueryArgs = Check.NotNull(queryArgs, nameof(queryArgs));
@@ -51,20 +51,35 @@ namespace Beef.Data.Cosmos
         }
 
         /// <summary>
-        /// Gets the <see cref="CosmosDbArgs"/>.
+        /// Gets the <see cref="CosmosDbArgs{T}"/>.
         /// </summary>
-        public CosmosDbArgs QueryArgs { get; private set; }
+        public CosmosDbArgs<T> QueryArgs { get; private set; }
 
         /// <summary>
         /// Manages the underlying query construction and lifetime.
         /// </summary>
         internal void ExecuteQuery(Action<IQueryable<T>> execute)
         {
-            CosmosDbInvoker.Default.Invoke(this, () =>
+            CosmosDbInvoker.Default.Invoke(this, () => ExecuteQueryInternal(execute), _db);
+        }
+
+        /// <summary>
+        /// Actually manage the underlying query construction and lifetime.
+        /// </summary>
+        private IQueryable<T> ExecuteQueryInternal(Action<IQueryable<T>> execute)
+        {
+            IQueryable<T> q = _db.GetContainer(QueryArgs.ContainerId).GetItemLinqQueryable<T>(allowSynchronousQueryExecution: true, requestOptions: _db.GetQueryRequestOptions(QueryArgs));
+            if (QueryArgs is ICosmosDbArgs qa && qa.IsTypeValue)
             {
-                var q = _db.GetContainer(QueryArgs.ContainerId).GetItemLinqQueryable<T>(allowSynchronousQueryExecution: true, requestOptions: _db.GetQueryRequestOptions(QueryArgs));
-                execute(_query == null ? Internal.CosmosDbHelper.AddTypeWhereClause<T>(q) : _query(Internal.CosmosDbHelper.AddTypeWhereClause<T>(q)));
-            }, _db);
+                q = _query == null ?
+                    Internal.CosmosDbHelper.AddTypeWhereClause<T>(q, qa.TypeValueType) :
+                    _query(Internal.CosmosDbHelper.AddTypeWhereClause<T>(q, qa.TypeValueType));
+            }
+            else
+                q = _query == null ? q : _query(q);
+
+            execute?.Invoke(q);
+            return q;
         }
 
         /// <summary>
@@ -74,21 +89,28 @@ namespace Beef.Data.Cosmos
         {
             return CosmosDbInvoker.Default.Invoke(this, () =>
             {
-                var q = _db.GetContainer(QueryArgs.ContainerId).GetItemLinqQueryable<T>(allowSynchronousQueryExecution: true, requestOptions: _db.GetQueryRequestOptions(QueryArgs));
-                return execute(_query == null ? Internal.CosmosDbHelper.AddTypeWhereClause<T>(q) : _query(Internal.CosmosDbHelper.AddTypeWhereClause<T>(q)));
+                return execute(AsQueryable(false));
             }, _db);
         }
 
         /// <summary>
-        /// Sets the paging from the <see cref="PagingArgs"/>.
+        /// Gets a prepared <see cref="IQueryable{T}"/> with any <see cref="CosmosDbTypeValue{T}"/> filtering as applicable.
         /// </summary>
-        internal IQueryable<T> SetPaging(IQueryable<T> query, PagingArgs paging)
+        /// <remarks>The <see cref="ICosmosDbArgs.Paging"/> is not supported.</remarks>
+        public IQueryable<T> AsQueryable()
         {
-            var q = query;
-            if (paging != null && paging.Skip > 0)
-                q = q.Skip((int)paging.Skip);
+            return AsQueryable(true);
+        }
 
-            return q.Take((int)(paging == null ? PagingArgs.DefaultTake : paging.Take));
+        /// <summary>
+        /// Initiate the IQueryable.
+        /// </summary>
+        private IQueryable<T> AsQueryable(bool checkPaging)
+        {
+            if (checkPaging && QueryArgs.Paging != null)
+                throw new NotSupportedException("The QueryArgs.Paging must be null for an AsQueryable(); this is a limitation of the Microsoft.Azure.Cosmos SDK in that the paging must be applied last, as such use the IQueryable.Paging provided to perform where appropriate.");
+
+            return ExecuteQueryInternal(null);
         }
 
         #region SelectSingle/SelectFirst
@@ -101,7 +123,7 @@ namespace Beef.Data.Cosmos
         {
             return ExecuteQuery(q =>
             {
-                q = SetPaging(q, PagingTop2);
+                q = q.Paging(0, 2);
                 return CosmosDbBase.GetAndFormatValue(q.AsEnumerable().Single());
             });
         }
@@ -114,7 +136,7 @@ namespace Beef.Data.Cosmos
         {
             return ExecuteQuery(q =>
             {
-                q = SetPaging(q, PagingTop2);
+                q = q.Paging(0, 2);
                 return CosmosDbBase.GetAndFormatValue(q.AsEnumerable().SingleOrDefault());
             });
         }
@@ -127,7 +149,7 @@ namespace Beef.Data.Cosmos
         {
             return ExecuteQuery(q =>
             {
-                q = SetPaging(q, PagingTop1);
+                q = q.Paging(0, 1);
                 return CosmosDbBase.GetAndFormatValue(q.AsEnumerable().First());
             });
         }
@@ -140,7 +162,7 @@ namespace Beef.Data.Cosmos
         {
             return ExecuteQuery(q =>
             {
-                q = SetPaging(q, PagingTop1);
+                q = q.Paging(0, 1);
                 return CosmosDbBase.GetAndFormatValue(q.AsEnumerable().FirstOrDefault());
             });
         }
@@ -154,6 +176,7 @@ namespace Beef.Data.Cosmos
         /// </summary>
         /// <typeparam name="TColl">The collection <see cref="Type"/>.</typeparam>
         /// <returns>A resultant collection.</returns>
+        /// <remarks>The <see cref="QueryArgs"/> <see cref="ICosmosDbArgs.Paging"/> is also applied, including <see cref="PagingArgs.IsGetCount"/> where requested.</remarks>
         public TColl SelectQuery<TColl>() where TColl : ICollection<T>, new()
         {
             var coll = new TColl();
@@ -166,11 +189,12 @@ namespace Beef.Data.Cosmos
         /// </summary>
         /// <typeparam name="TColl">The collection <see cref="Type"/>.</typeparam>
         /// <param name="coll">The collection to add items to.</param>
+        /// <remarks>The <see cref="QueryArgs"/> <see cref="ICosmosDbArgs.Paging"/> is also applied, including <see cref="PagingArgs.IsGetCount"/> where requested.</remarks>
         public void SelectQuery<TColl>(TColl coll) where TColl : ICollection<T>
         {
             ExecuteQuery(query => 
             {
-                foreach (var item in SetPaging(query, QueryArgs.Paging).AsEnumerable())
+                foreach (var item in query.Paging(QueryArgs.Paging).AsEnumerable())
                 {
                     coll.Add(CosmosDbBase.GetAndFormatValue(item));
                 }